implement HarperDBBase class
