+++ harperdb/wrappers.py
@@ -1,13 +1,10 @@
-import base64
 import csv
 import datetime
-import json
-import requests
 
-from .exceptions import *
+from .harperdb_base import *
 
 
-class HarperDBWrapper():
+class HarperDBWrapper(HarperDBBase):
 
     """ HarperDBWrapper provides a high-level, object-oriented interface for
     HarperDB. From this top-level object an application programmer can make
@@ -89,12 +86,12 @@ class HarperDBWrapper():
 
     def __iter__(self):
         # get a current dict of schemas and iterate over that
-        return _HarperDBSchemas(self._get_schemas())
+        return _HarperDBSchemas(self.__get_schemas())
 
     def __len__(self):
-        return len(self._get_schemas())
+        return len(self.__get_schemas())
 
-    def _get_schemas(self):
+    def __get_schemas(self):
         """ Returns a dictionary of schemas, subscriptable by name.
         """
         return_value = dict()
@@ -103,31 +100,6 @@ class HarperDBWrapper():
             schema_object = HarperDBSchema(schema, self)
             return_value[schema] = schema_object
         return return_value
-
-    def __make_request(self, data):
-        """ Make a POST request to the database instance with JSON data.
-
-        Returns JSON response, raises HarperDBError if the server returns 500.
-        """
-        headers = {
-            'Content-Type': 'application/json',
-        }
-        if self.token:
-            headers['Authorization'] = self.token
-        response = requests.request(
-            'POST',
-            self.url,
-            headers=headers,
-            data=json.dumps(data),
-            timeout=self.timeout)
-        body = response.json()
-        try:
-            response.raise_for_status()
-        except requests.exceptions.HTTPError:
-            raise HarperDBError(body.get('error', 'An unknown error occurred'))
-        return body
-
-    # High-Level Methods
 
     def create_schema(self, name):
         """ Create a schema in this database.
@@ -139,171 +111,6 @@ class HarperDBWrapper():
         """ Drop a schema from this database.
         """
         self._drop_schema(name)
-
-    # Low-Level Methods
-    # Schemas and Tables
-
-    def _create_schema(self, schema):
-        return self.__make_request({
-            'operation': 'create_schema',
-            'schema': schema,
-        })
-
-    def _drop_schema(self, schema):
-        return self.__make_request({
-            'operation': 'drop_schema',
-            'schema': schema,
-        })
-
-    def _describe_schema(self, schema):
-        return self.__make_request({
-            'operation': 'describe_schema',
-            'schema': schema,
-        })
-
-    def _create_table(self, schema, table, hash_attribute):
-        return self.__make_request({
-            'operation': 'create_table',
-            'schema': schema,
-            'table': table,
-            'hash_attribute': hash_attribute,
-        })
-
-    def _describe_table(self, schema, table):
-        return self.__make_request({
-            'operation': 'describe_table',
-            'schema': schema,
-            'table': table,
-        })
-
-    def _describe_all(self):
-        return self.__make_request({
-            'operation': 'describe_all',
-        })
-
-    def _drop_table(self, schema, table):
-        return self.__make_request({
-            'operation': 'drop_table',
-            'schema': schema,
-            'table': table,
-        })
-
-    def _drop_attribute(self, schema, table, attribute):
-        return self.__make_request({
-            'operation': 'drop_attribute',
-            'schema': schema,
-            'table': table,
-            'attribute': attribute,
-        })
-
-    # Low-Level Methods
-    # NoSQL Operations
-
-    def _insert(self, schema, table, records):
-        return self.__make_request({
-            'operation': 'insert',
-            'schema': schema,
-            'table': table,
-            'records': records,
-        })
-
-    def _update(self, schema, table, records):
-        return self.__make_request({
-            'operation': 'update',
-            'schema': schema,
-            'table': table,
-            'records': records,
-        })
-
-    def _delete(self, schema, table, hash_values):
-        return self.__make_request({
-            'operation': 'delete',
-            'schema': schema,
-            'table': table,
-            'hash_values': hash_values,
-        })
-
-    def _search_by_hash(
-            self,
-            schema,
-            table,
-            hash_values,
-            get_attributes=['*']):
-        return self.__make_request({
-            'operation': 'search_by_hash',
-            'schema': schema,
-            'table': table,
-            'hash_values': hash_values,
-            'get_attributes': get_attributes,
-        })
-
-    def _search_by_value(
-            self,
-            schema,
-            table,
-            search_attribute,
-            search_value,
-            get_attributes=['*']):
-        return self.__make_request({
-            'operation': 'search_by_value',
-            'schema': schema,
-            'table': table,
-            'search_attribute': search_attribute,
-            'search_value': search_value,
-            'get_attributes': get_attributes,
-        })
-
-    # Low-Level Methods
-    # SQL Operations
-
-    def _sql(self, sql_string):
-        return self.__make_request({
-            'operation': 'sql',
-            'sql': sql_string,
-        })
-
-    # Low-Level Methods
-    # CSV Operations
-
-    def _csv_data_load(self, schema, table, path, action='insert'):
-        with open(path) as csv_file:
-            data = csv_file.read()
-        return self.__make_request({
-            'operation': 'csv_data_load',
-            'action': action,
-            'schema': schema,
-            'table': table,
-            'data': data,
-        })
-
-
-    def _csv_file_load(self, schema, table, file_path, action='insert'):
-        return self.__make_request({
-            'operation': 'csv_file_load',
-            'action': action,
-            'schema': schema,
-            'table': table,
-            'file_path': file_path,
-        })
-
-    def _csv_url_load(self, schema, table, csv_url, action='insert'):
-        return self.__make_request({
-            'operation': 'csv_url_load',
-            'action': action,
-            'schema': schema,
-            'table': table,
-            'csv_url': csv_url,
-        })
-
-    # Low-Level Methods
-    # Jobs
-
-    def _get_job(self, id):
-        return self.__make_request({
-            'operation': 'get_job',
-            'id': id,
-        })
-
 
 
 class HarperDBRecord():
